+++ src/gx/GXFrameBuf.c
@@ -409,9 +409,9 @@ GXRenderModeObj GXEurgb60Hz480IntDf = {
  */
 void GXAdjustForOverscan(GXRenderModeObj* rmin, GXRenderModeObj* rmout, u16 hor, u16 ver)
 {
-	unsigned short hor2 = hor * 2;
-	unsigned short ver2 = ver * 2;
-	unsigned long verf;
+	u16 hor2 = hor * 2;
+	u16 ver2 = ver * 2;
+	u32 verf;
 
 	if (rmin != rmout) {
 		*rmout = *rmin;
